bigint -> int

+++ backend/app/Models/Order.php
@@ -10,7 +10,7 @@ class Order extends Model
     protected $table = 'orders';
     protected $primaryKey = 'id';
     public $incrementing = false;
-    protected $keyType = 'bigint';
+    protected $keyType = 'int';
 
     protected $fillable = [
         'id',
